+++ python-runnables/inspect-node-pools/runnable.py
@@ -30,57 +30,34 @@ class MyRunnable(Runnable):
         connection_info = get_connection_info(dss_cluster_config.get('config'))
         
         node_group_id = self.config.get('nodeGroupId', None)
-        
-        if node_group_id is None or len(node_group_id) == 0:
-            args = ['get', 'nodegroup']
-            #args = args + ['-v', '4']
-            args = args + ['--cluster', cluster_id]
-            args = args + get_region_arg(connection_info)
-            args = args + ['-o', 'json']
 
-            c = EksctlCommand(args, connection_info)
-            node_groups = json.loads(c.run_and_get_output())
-            node_group_ids = [node_group['Name'] for node_group in node_groups]
-        else:
-            node_group_ids = [node_group_id]
+        # There's a bug in `eksctl` where the `StackName` doesn't appear in the output
+        # if `get nodegroup` is called with a single node group name and that node
+        # group is "managed". So, always get all node groups, even if only one is
+        # specified in the params and even though we don't use the `StackName` anymore.
+        args = ['get', 'nodegroup']
+        args = args + ['--cluster', cluster_id]
+        args = args + get_region_arg(connection_info)
+        args = args + ['-o', 'json']
 
-        node_groups = []
-        for node_group_id in node_group_ids:
-            args = ['get', 'nodegroup']
-            #args = args + ['-v', '4']
-            args = args + ['--cluster', cluster_id]
-            args = args + ['--name', node_group_id]
-            args = args + get_region_arg(connection_info)
-            args = args + ['-o', 'json']
+        c = EksctlCommand(args, connection_info)
+        node_groups = json.loads(c.run_and_get_output())
 
-            c = EksctlCommand(args, connection_info)
-            node_group_batch = json.loads(c.run_and_get_output())
-            if len(node_group_batch) == 0:
-                node_groups.append('<h5>%s</h5><div class="alert alert-error">Unable to get details</div>' % (node_group_id))
-                continue
-                
-            node_group = node_group_batch[0]
+        if node_group_id and len(node_group_id) != 0:
+            node_groups = list(filter(lambda node_group: node_group.get('Name') == node_group_id, node_groups))
 
-            node_group_stack_name = node_group['StackName']
-            
-            args = ['cloudformation', 'describe-stack-resources']
-            args = args + ['--stack-name', node_group_stack_name]
+            if len(node_groups) == 0:
+                return '<div><h5>%s</h5><div class="alert alert-error">Unable to get details</div></div>' % (node_group_id)
 
-            c = AwsCommand(args, connection_info)
-            node_group_resources = json.loads(c.run_and_get_output()).get('StackResources', [])
-            
-            # find the auto-scaling-group
-            auto_scaling_resource = None
-            for r in node_group_resources:
-                if r.get('ResourceType', '') == 'AWS::AutoScaling::AutoScalingGroup':
-                    auto_scaling_resource = r
-                
-            if auto_scaling_resource is None:
-                node_groups.append('<h5>%s</h5><div class="alert alert-error">Unable to get auto-scaling group</div><pre class="debug">%s</pre>' % (node_group_id, json.dumps(node_group, indent=2)))
+        node_group_outputs = []
+        for node_group in node_groups:
+            node_group_id = node_group.get('Name')
+            node_group_auto_scaling_id = node_group.get('AutoScalingGroupName')
+
+            if node_group_auto_scaling_id is None:
+                node_group_outputs.append('<h5>%s</h5><div class="alert alert-error">Unable to get auto-scaling group</div><pre class="debug">%s</pre>' % (node_group_id, json.dumps(node_group, indent=2)))
                 continue
 
-            node_group_auto_scaling_id = auto_scaling_resource['PhysicalResourceId']
-            
             args = ['autoscaling', 'describe-auto-scaling-groups']
             args = args + ['--auto-scaling-group-names', node_group_auto_scaling_id]
 
@@ -88,7 +65,7 @@ class MyRunnable(Runnable):
             auto_scaling_resources = json.loads(c.run_and_get_output()).get('AutoScalingGroups', [])
             
             if len(auto_scaling_resources) == 0:
-                node_groups.append('<h5>%s</h5><div class="alert alert-error">Unable to get auto-scaling group\'s resources</div><pre class="debug">%s</pre>' % (node_group_id, json.dumps(node_group, indent=2)))
+                node_group_outputs.append('<h5>%s</h5><div class="alert alert-error">Unable to get auto-scaling group\'s resources</div><pre class="debug">%s</pre>' % (node_group_id, json.dumps(node_group, indent=2)))
                 continue
                 
             auto_scaling_resource = auto_scaling_resources[0]
@@ -96,6 +73,6 @@ class MyRunnable(Runnable):
             min_instances = auto_scaling_resource.get('MinSize','')
             cur_instances = len(auto_scaling_resource.get('Instances',[]))
             max_instances = auto_scaling_resource.get('MaxSize','')
-            node_groups.append('<h5>%s</h5><pre class="debug">%s</pre><div>Min=%s, current=%s, max=%s</div><pre class="debug">%s</pre>' % (node_group_id, json.dumps(node_group, indent=2), min_instances, cur_instances, max_instances, json.dumps(auto_scaling_resource.get('Instances', []), indent=2)))
+            node_group_outputs.append('<h5>%s</h5><pre class="debug">%s</pre><div>Min=%s, current=%s, max=%s</div><pre class="debug">%s</pre>' % (node_group_id, json.dumps(node_group, indent=2), min_instances, cur_instances, max_instances, json.dumps(auto_scaling_resource.get('Instances', []), indent=2)))
         
-        return '<div>%s</div>' % ''.join(node_groups)
+        return '<div>%s</div>' % ''.join(node_group_outputs)